requirements.txt
instaloader>=4.13
openai>=1.30,<2.0
google-api-python-client>=2.130
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
requests>=2.31
python-dotenv>=1.0
openpyxl>=3.1
